eventia-backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.5.0
marshmallow==3.20.1
prometheus-client==0.17.1
psutil==5.9.5
waitress==2.1.2
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.3
pytest-cov==4.1.0
pytest-mock==3.11.1
requests==2.31.0
Pillow==10.0.0
aiofiles==23.2.1
motor==3.2.0
pydantic==1.10.8